Provide debug info via the about dialog
No longer need a separate error-reports service.
