--- gaphor/plugins/errorreports/__init__.py
@@ -1,3 +1,0 @@
-# ruff: noqa: F401
-
-from gaphor.plugins.errorreports.errorreports import ErrorReports
