--- gaphor/plugins/errorreports/errorreports.ui
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<interface>
-  <requires lib="gtk" version="4.0"/>
-  <object class="GtkTextBuffer" id="buffer"/>
-  <object class="GtkWindow" id="error-reports">
-    <property name="title" translatable="yes">Error Reports</property>
-    <property name="default-width">600</property>
-    <property name="default-height">400</property>
-    <child type="titlebar">
-      <object class="AdwHeaderBar">
-        <property name="decoration-layout">:close</property>
-      </object>
-    </child>
-    <child>
-      <object class="GtkBox">
-        <property name="orientation">vertical</property>
-        <child>
-          <object class="GtkLabel">
-            <property name="margin-start">6</property>
-            <property name="margin-end">6</property>
-            <property name="margin-top">6</property>
-            <property name="margin-bottom">6</property>
-            <property name="label" translatable="yes">Any unhandled errors are shown here. You can use these to provide extra information when you create an issue on GitHub.</property>
-            <property name="wrap">1</property>
-            <property name="xalign">0</property>
-          </object>
-        </child>
-        <child>
-          <object class="GtkScrolledWindow">
-            <property name="vexpand">1</property>
-            <property name="focusable">1</property>
-            <property name="child">
-              <object class="GtkTextView">
-                <property name="focusable">1</property>
-                <property name="editable">0</property>
-                <property name="buffer">buffer</property>
-                <property name="monospace">1</property>
-              </object>
-            </property>
-          </object>
-        </child>
-      </object>
-    </child>
-  </object>
-</interface>
